Replace DefaultCondition by DefaultFunction
