+++ core/src/main/java/io/doov/core/dsl/field/types/DoubleFieldInfo.java
@@ -15,6 +15,9 @@
  */
 package io.doov.core.dsl.field.types;
 
+import static io.doov.core.dsl.impl.DefaultFunction.valueModel;
+import static io.doov.core.dsl.meta.predicate.FieldMetadata.fieldMetadata;
+
 import io.doov.core.FieldInfo;
 import io.doov.core.dsl.field.DelegatingFieldInfoImpl;
 import io.doov.core.dsl.impl.num.DoubleFunction;
@@ -27,7 +30,7 @@ public class DoubleFieldInfo extends DelegatingFieldInfoImpl implements NumericF
 
     @Override
     public DoubleFunction getNumericFunction() {
-        return new DoubleFunction(this);
+        return new DoubleFunction(fieldMetadata(this), (m, c) -> valueModel(m, this));
     }
 
 }
