+++ core/src/main/java/io/doov/core/dsl/field/types/IntegerFieldInfo.java
@@ -15,6 +15,9 @@
  */
 package io.doov.core.dsl.field.types;
 
+import static io.doov.core.dsl.impl.DefaultFunction.valueModel;
+import static io.doov.core.dsl.meta.predicate.FieldMetadata.fieldMetadata;
+
 import io.doov.core.FieldInfo;
 import io.doov.core.dsl.field.DelegatingFieldInfoImpl;
 import io.doov.core.dsl.impl.num.IntegerFunction;
@@ -27,7 +30,7 @@ public class IntegerFieldInfo extends DelegatingFieldInfoImpl implements Numeric
 
     @Override
     public IntegerFunction getNumericFunction() {
-        return new IntegerFunction(this);
+        return new IntegerFunction(fieldMetadata(this), (m, c) -> valueModel(m, this));
     }
 
 }
