+++ core/src/main/java/io/doov/core/dsl/field/types/LongFieldInfo.java
@@ -15,6 +15,9 @@
  */
 package io.doov.core.dsl.field.types;
 
+import static io.doov.core.dsl.impl.DefaultFunction.valueModel;
+import static io.doov.core.dsl.meta.predicate.FieldMetadata.fieldMetadata;
+
 import io.doov.core.FieldInfo;
 import io.doov.core.dsl.field.DelegatingFieldInfoImpl;
 import io.doov.core.dsl.impl.num.LongFunction;
@@ -27,7 +30,7 @@ public class LongFieldInfo extends DelegatingFieldInfoImpl implements NumericFie
 
     @Override
     public LongFunction getNumericFunction() {
-        return new LongFunction(this);
+        return new LongFunction(fieldMetadata(this), (m, c) -> valueModel(m, this));
     }
 
 }
