+++ core/src/main/java/io/doov/core/dsl/field/types/StringFieldInfo.java
@@ -15,6 +15,9 @@
  */
 package io.doov.core.dsl.field.types;
 
+import static io.doov.core.dsl.impl.DefaultFunction.valueModel;
+import static io.doov.core.dsl.meta.predicate.FieldMetadata.fieldMetadata;
+
 import io.doov.core.FieldInfo;
 import io.doov.core.dsl.field.DelegatingFieldInfoImpl;
 import io.doov.core.dsl.impl.base.StringFunction;
@@ -27,7 +30,7 @@ public class StringFieldInfo extends DelegatingFieldInfoImpl implements TextFiel
 
     @Override
     public StringFunction getStringFunction() {
-        return new StringFunction(this);
+        return new StringFunction(fieldMetadata(this), (m, c) -> valueModel(m, this));
     }
 
 }
