+++ core/src/main/java/io/doov/core/dsl/impl/num/DoubleFunction.java
@@ -19,15 +19,10 @@ import java.util.Optional;
 import java.util.function.BiFunction;
 
 import io.doov.core.FieldModel;
-import io.doov.core.dsl.DslField;
 import io.doov.core.dsl.lang.Context;
 import io.doov.core.dsl.meta.predicate.PredicateMetadata;
 
 public class DoubleFunction extends NumericFunction<Double> implements DoubleOperators {
-
-    public DoubleFunction(DslField<Double> field) {
-        super(field);
-    }
 
     public DoubleFunction(PredicateMetadata metadata, BiFunction<FieldModel, Context, Optional<Double>> value) {
         super(metadata, value);
