+++ core/src/main/java/io/doov/core/dsl/impl/time/LocalDateFunction.java
@@ -20,15 +20,10 @@ import java.util.Optional;
 import java.util.function.BiFunction;
 
 import io.doov.core.FieldModel;
-import io.doov.core.dsl.DslField;
 import io.doov.core.dsl.lang.Context;
 import io.doov.core.dsl.meta.predicate.PredicateMetadata;
 
 public class LocalDateFunction extends TemporalFunction<LocalDate> implements LocalDateOperators {
-
-    public LocalDateFunction(DslField<LocalDate> field) {
-        super(field);
-    }
 
     public LocalDateFunction(PredicateMetadata metadata, BiFunction<FieldModel, Context, Optional<LocalDate>> value) {
         super(metadata, value);
